requirements.txt
Django==2.2.6
django-appconf==1.0.3
django-bootstrap-modal-forms==1.4.4
django-searchable-select==1.5.0
django-select2==7.1.1
django-widget-tweaks==1.4.5
isort==4.3.21
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.4
Pillow==6.2.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
